FVI/requirements.txt
# FVI Phase 1 Dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1

# Configuration and utilities
PyYAML>=6.0
python-dateutil>=2.8.0

# AI/ML libraries
google-generativeai>=0.3.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Logging and utilities
tqdm>=4.65.0
